# Improve performance of String.Equals(..., OrdinalIgnoreCase)