--- a/src/System.Private.CoreLib/shared/System/Globalization/CompareInfo.cs
+++ b/src/System.Private.CoreLib/shared/System/Globalization/CompareInfo.cs
@@ -16,7 +16,7 @@ using System.Reflection;
 using System.Diagnostics;
 using System.Runtime.InteropServices;
 using System.Runtime.Serialization;
-using System.Buffers;
+using System.Text;
 using Internal.Runtime.CompilerServices;
 
 namespace System.Globalization
@@ -595,37 +595,87 @@ namespace System.Globalization
         }
 
 
-        internal static bool EqualsOrdinalIgnoreCase(ref char strA, ref char strB, int length)
+        internal static bool EqualsOrdinalIgnoreCase(ref char charA, ref char charB, int length)
         {
-            ref char charA = ref strA;
-            ref char charB = ref strB;
-
-            // in InvariantMode we support all range and not only the ascii characters.
-            char maxChar = (GlobalizationMode.Invariant ? (char)0xFFFF : (char)0x7F);
-
-            while (length != 0 && charA <= maxChar && charB <= maxChar)
+            if (!GlobalizationMode.Invariant)
             {
-                // Ordinal equals or lowercase equals if the result ends up in the a-z range 
-                if (charA == charB ||
-                    ((charA | 0x20) == (charB | 0x20) &&
-                        (uint)((charA | 0x20) - 'a') <= (uint)('z' - 'a')))
+                IntPtr byteOffset = IntPtr.Zero;
+                while ((uint)length >= 2)
                 {
-                    length--;
-                    charA = ref Unsafe.Add(ref charA, 1);
-                    charB = ref Unsafe.Add(ref charB, 1);
-                }
-                else
-                {
-                    return false;
-                }
-            }
+                    // Read 2 chars (32 bits) at a time from each string
+                    uint valueA = Unsafe.ReadUnaligned<uint>(ref Unsafe.As<char, byte>(ref Unsafe.AddByteOffset(ref charA, byteOffset)));
+                    uint valueB = Unsafe.ReadUnaligned<uint>(ref Unsafe.As<char, byte>(ref Unsafe.AddByteOffset(ref charB, byteOffset)));
 
-            if (length == 0)
+                    if (!Utf16Utility.AllCharsInUInt32AreAscii(valueA | valueB))
+                    {
+                        goto NonAscii; // one of the inputs contains non-ASCII data
+                    }
+
+                    // Generally, the caller has likely performed a first-pass check that the input strings
+                    // are likely equal. Consider a dictionary which computes the hash code of its key before
+                    // performing a proper deep equality check of the string contents. We want to optimize for
+                    // the case where the equality check is likely to succeed, which means that we want to avoid
+                    // branching within this loop unless we're about to exit the loop, either due to failure or
+                    // due to us running out of input data.
+
+                    if (!Utf16Utility.UInt32OrdinalIgnoreCaseAscii(valueA, valueB))
+                    {
+                        return false;
+                    }
+
+                    byteOffset += 2;
+                    length -= 2;
+                }
+
+                if (length != 0)
+                {
+                    Debug.Assert(length == 1);
+
+                    uint valueA = Unsafe.AddByteOffset(ref charA, byteOffset);
+                    uint valueB = Unsafe.AddByteOffset(ref charB, byteOffset);
+
+                    if ((valueA | valueB) > 0x7Fu)
+                    {
+                        goto NonAscii; // one of the inputs contains non-ASCII data
+                    }
+
+                    if (valueA == valueB)
+                    {
+                        return true;
+                    }
+
+                    valueA |= 0x20u;
+                    return ((uint)(valueA - 'a') <= (uint)('z' - 'a'))
+                        && (valueA == (valueB | 0x20u));
+                }
+
+                Debug.Assert(length == 0);
                 return true;
 
-            Debug.Assert(!GlobalizationMode.Invariant);
+            NonAscii:
+                return CompareStringOrdinalIgnoreCase(ref Unsafe.AddByteOffset(ref charA, byteOffset), length, ref Unsafe.AddByteOffset(ref charB, byteOffset), length) == 0;
+            }
+            else
+            {
+                while (length != 0)
+                {
+                    // Ordinal equals or lowercase equals if the result ends up in the a-z range 
+                    if (charA == charB ||
+                        ((charA | 0x20) == (charB | 0x20) &&
+                            (uint)((charA | 0x20) - 'a') <= (uint)('z' - 'a')))
+                    {
+                        length--;
+                        charA = ref Unsafe.Add(ref charA, 1);
+                        charB = ref Unsafe.Add(ref charB, 1);
+                    }
+                    else
+                    {
+                        return false;
+                    }
+                }
 
-            return CompareStringOrdinalIgnoreCase(ref charA, length, ref charB, length) == 0;
+                return true;
+            }
         }
 
         ////////////////////////////////////////////////////////////////////////