--- a/src/System.Private.CoreLib/shared/System/StringComparer.cs
+++ b/src/System.Private.CoreLib/shared/System/StringComparer.cs
@@ -379,6 +379,11 @@ namespace System
                 return false;
             }
 
+            if (x.Length != y.Length)
+            {
+                return false;
+            }
+
             return CompareInfo.EqualsOrdinalIgnoreCase(ref x.GetRawStringData(), ref y.GetRawStringData(), x.Length);
         }