This improves the throughput of `OrdinalIgnoreCase` string comparisons, especially for smaller inputs.

|                                          Method |      Toolchain | StringLength |      Mean |     Error |    StdDev | Scaled | ScaledSD |
|------------------------------------------------ |--------------- |------------- |----------:|----------:|----------:|-------:|---------:|
|     **EqualsOrdinalIgnoreCase_StringsAreIdentical** |   **experimental** |            **1** |  **4.983 ms** | **0.0603 ms** | **0.0564 ms** |   **0.51** |     **0.01** |
|     EqualsOrdinalIgnoreCase_StringsAreIdentical | preview1-27030 |            1 |  9.845 ms | 0.0658 ms | 0.0550 ms |   1.00 |     0.00 |
|                                                 |                |              |           |           |           |        |          |
| EqualsOrdinalIgnoreCase_StringsDifferOnlyInCase |   experimental |            1 |  5.556 ms | 0.0759 ms | 0.0673 ms |   0.63 |     0.01 |
| EqualsOrdinalIgnoreCase_StringsDifferOnlyInCase | preview1-27030 |            1 |  8.794 ms | 0.0832 ms | 0.0738 ms |   1.00 |     0.00 |
|                                                 |                |              |           |           |           |        |          |
|     **EqualsOrdinalIgnoreCase_StringsAreIdentical** |   **experimental** |            **2** |  **5.520 ms** | **0.0751 ms** | **0.0627 ms** |   **0.53** |     **0.01** |
|     EqualsOrdinalIgnoreCase_StringsAreIdentical | preview1-27030 |            2 | 10.395 ms | 0.2048 ms | 0.2664 ms |   1.00 |     0.00 |
|                                                 |                |              |           |           |           |        |          |
| EqualsOrdinalIgnoreCase_StringsDifferOnlyInCase |   experimental |            2 |  5.498 ms | 0.0679 ms | 0.0602 ms |   0.56 |     0.01 |
| EqualsOrdinalIgnoreCase_StringsDifferOnlyInCase | preview1-27030 |            2 |  9.824 ms | 0.1596 ms | 0.1415 ms |   1.00 |     0.00 |
|                                                 |                |              |           |           |           |        |          |
|     **EqualsOrdinalIgnoreCase_StringsAreIdentical** |   **experimental** |            **3** |  **6.252 ms** | **0.0609 ms** | **0.0539 ms** |   **0.59** |     **0.01** |
|     EqualsOrdinalIgnoreCase_StringsAreIdentical | preview1-27030 |            3 | 10.665 ms | 0.2004 ms | 0.1968 ms |   1.00 |     0.00 |
|                                                 |                |              |           |           |           |        |          |
| EqualsOrdinalIgnoreCase_StringsDifferOnlyInCase |   experimental |            3 |  6.674 ms | 0.0549 ms | 0.0487 ms |   0.56 |     0.02 |
| EqualsOrdinalIgnoreCase_StringsDifferOnlyInCase | preview1-27030 |            3 | 11.842 ms | 0.2349 ms | 0.4903 ms |   1.00 |     0.00 |
|                                                 |                |              |           |           |           |        |          |
|     **EqualsOrdinalIgnoreCase_StringsAreIdentical** |   **experimental** |            **4** |  **7.526 ms** | **0.1498 ms** | **0.2776 ms** |   **0.66** |     **0.02** |
|     EqualsOrdinalIgnoreCase_StringsAreIdentical | preview1-27030 |            4 | 11.392 ms | 0.1014 ms | 0.0847 ms |   1.00 |     0.00 |
|                                                 |                |              |           |           |           |        |          |
| EqualsOrdinalIgnoreCase_StringsDifferOnlyInCase |   experimental |            4 |  7.659 ms | 0.1812 ms | 0.5343 ms |   0.66 |     0.05 |
| EqualsOrdinalIgnoreCase_StringsDifferOnlyInCase | preview1-27030 |            4 | 11.666 ms | 0.1836 ms | 0.1433 ms |   1.00 |     0.00 |
|                                                 |                |              |           |           |           |        |          |
|     **EqualsOrdinalIgnoreCase_StringsAreIdentical** |   **experimental** |            **5** |  **7.247 ms** | **0.0681 ms** | **0.0637 ms** |   **0.62** |     **0.01** |
|     EqualsOrdinalIgnoreCase_StringsAreIdentical | preview1-27030 |            5 | 11.668 ms | 0.1598 ms | 0.1495 ms |   1.00 |     0.00 |
|                                                 |                |              |           |           |           |        |          |
| EqualsOrdinalIgnoreCase_StringsDifferOnlyInCase |   experimental |            5 |  7.850 ms | 0.1320 ms | 0.1234 ms |   0.65 |     0.01 |
| EqualsOrdinalIgnoreCase_StringsDifferOnlyInCase | preview1-27030 |            5 | 12.053 ms | 0.1226 ms | 0.1147 ms |   1.00 |     0.00 |
|                                                 |                |              |           |           |           |        |          |
|     **EqualsOrdinalIgnoreCase_StringsAreIdentical** |   **experimental** |            **8** |  **9.169 ms** | **0.1456 ms** | **0.1362 ms** |   **0.60** |     **0.01** |
|     EqualsOrdinalIgnoreCase_StringsAreIdentical | preview1-27030 |            8 | 15.190 ms | 0.1261 ms | 0.1117 ms |   1.00 |     0.00 |
|                                                 |                |              |           |           |           |        |          |
| EqualsOrdinalIgnoreCase_StringsDifferOnlyInCase |   experimental |            8 |  9.318 ms | 0.0796 ms | 0.0621 ms |   0.62 |     0.01 |
| EqualsOrdinalIgnoreCase_StringsDifferOnlyInCase | preview1-27030 |            8 | 15.073 ms | 0.1343 ms | 0.1256 ms |   1.00 |     0.00 |
|                                                 |                |              |           |           |           |        |          |
|     **EqualsOrdinalIgnoreCase_StringsAreIdentical** |   **experimental** |           **12** | **12.115 ms** | **0.2605 ms** | **0.2309 ms** |   **0.64** |     **0.02** |
|     EqualsOrdinalIgnoreCase_StringsAreIdentical | preview1-27030 |           12 | 18.960 ms | 0.3438 ms | 0.3216 ms |   1.00 |     0.00 |
|                                                 |                |              |           |           |           |        |          |
| EqualsOrdinalIgnoreCase_StringsDifferOnlyInCase |   experimental |           12 | 12.761 ms | 0.2363 ms | 0.2095 ms |   0.64 |     0.01 |
| EqualsOrdinalIgnoreCase_StringsDifferOnlyInCase | preview1-27030 |           12 | 19.848 ms | 0.2431 ms | 0.2155 ms |   1.00 |     0.00 |

The performance gains come primarily from two changes. First, the inner loop is now checking 2 characters at a time instead of one character at a time. Second, the internal branching within the inner loop has been eliminated, which helps avoid branch mispredictions when mixed-case input is provided. The source code comment in the inner loop explains this in more detail.

This code isn't vectorized because strings being compared for equality tend to be small - dictionary keys, identifiers, and the like. I also experimented with performing operations on 4 chars (64 bits) at a time on 64-bit platforms. It provides for inputs of 4+ chars an approx. 15% throughput boost above and beyond what this PR provides. Ultimately I didn't include the 64-bit code because I thought it would complicate the method for too little gain. If consensus is that the gain is worth it I can reintroduce the 64-bit-specific code paths.