@GrabYourPitchforks 2018-11-01
I'll investigate the test failures. It's possible I ran the tests locally against the wrong version of the DLL.