@danmosemsft 2018-11-01
Given the test failures, apparently I overlooked a mistake.