@GrabYourPitchforks 2018-11-01
I'll update the main description with the benchmark results from the latest iteration shortly.