@GrabYourPitchforks 2018-11-01
Benchmark is running overnight and I will post updated results tomorrow.